Fix remaining issue with long kmers
This commit further fixes an issue that was partially fixed in the previous version. In places like the kmer seed table, uint8_t's were still used for kmer length and position variables. Now size_t's are used for all length and positions, meaning that all lengths of kmers will be supported.

+++ src/AwFmIndex.h
@@ -275,7 +275,7 @@ enum AwFmReturnCode awFmReadIndexFromFile(
  *    the given kmer does not exist in the database sequence.
  */
 struct AwFmSearchRange awFmFindSearchRangeForString(
-		const struct AwFmIndex *_RESTRICT_ const index, const char *_RESTRICT_ const kmer, const uint16_t kmerLength);
+		const struct AwFmIndex *_RESTRICT_ const index, const char *_RESTRICT_ const kmer, const size_t kmerLength);
 
 
 /*
